Initial commit
This sets up the initial skeleton, as well as a fully functioning RGB to
XYZ conversion. The matrix testing format should be a good pattern to
copy for all following tests.

RGB allows both hex scale [0,255] and [0,1] normalized values, and
requires a `normalized` bool be provided. This could certainly be
expanded in the future to a system where users could provide and
arbitrary scale, but does not seem necessary now.